pyproject.toml
[project]
name = "docklift"
description = "Deploy web applications to VPS using Docker Compose with automatic SSL"
readme = "README.md"
authors = [
    { name = "Amir Karimi", email = "me@amirkarimi.dev" }
]
requires-python = ">=3.12"
dependencies = [
    "fabric>=3.2.2",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "click>=8.1.7",
]
dynamic = ["version"]


[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
docklift = "docklift:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true
exclude = [
    "tests/",
    "build/",
]
